Remove unnecessary findProgramItems calls
