+++ server/src/features/program-item/utils/addSignupQuestions.ts
@@ -17,6 +17,7 @@ export const addSignupQuestions = async (): Promise<void> => {
   });
   await Promise.all(questionPromises);
 
+  // Most tournaments have same private signup question where attendee contact details are asked
   if (!tournamentSignupQuestion) {
     return;
   }
